feat: 实现端到端加密API、媒体API、语音消息API和VoIP API
主要更新:
- 修复 keys/upload 数据库约束问题
- 实现签名上传端点 (keys/signatures/upload)
- 实现设备签名密钥上传端点 (keys/device_signing/upload)
- 实现媒体下载指定文件名端点
- 修复媒体上传权限问题
- 修复语音消息数据库类型不匹配问题
- 实现完整的VoIP TURN凭证生成服务
- 修复房间状态API认证问题
- 修复时间戳转事件端点参数解析
- 更新api-error.md文档标记所有问题已修复
- 优化单元测试使用现有数据库表结构

+++ src/web/routes/voice.rs
@@ -321,7 +321,7 @@ async fn upload_voice_message(
     let duration_ms = body
         .get("duration_ms")
         .and_then(|v| v.as_i64())
-        .unwrap_or(0) as i32;
+        .unwrap_or(0) as i64;
 
     if duration_ms <= 0 {
         return Err(ApiError::bad_request(
